refactor(api): split chat retrieval boundary (#414)
